reorganized the features folder
